oscar q2_0 : VEC kernel fixes + HP scaffolding (WIP, not yet coherent)
Fix three nthreads_V==WARP_SIZE edge-case bugs in fattn-vec.cuh:
  - k position formula: distribute 128 threads across 4 groups of 32
  - VKQ_tmp shared-memory offset: per-group region, no overlap
  - i_VKQ index formula: bounds element indices within head dim
Override vec_dot template param to nthreads_V for q2_0 K type.
Add explicit casts in q2_0 vec_dot (integer-division + truncation bugs).
CPU type traits (vec_dot, clamp, arch-fallback) for link-clean build.
HP sink+recent buffer scaffolding in llama-kv-cache (compiles, not wired).
Template instances for D=64/128/256/512 q2_0 combos.

Build: clean (229/229 targets).
Coherence: still garbage at c=512. Root cause under investigation.

+++ ggml/src/ggml-cuda/fattn-common.cuh
@@ -153,7 +153,7 @@ static __device__ __forceinline__ float vec_dot_fattn_vec_KQ_q2_0(
         const int   u  = Q_q8[k_KQ_0/nthreads];
         const float2 Q_ds = ((const float2 *) Q_ds_v)[k_KQ_0/nthreads];
         const float q8scale = Q_ds.x;
-        const float q8off   = Q_ds.y / QI8_1;
+        const float q8off   = Q_ds.y / (float)QI8_1;
         #pragma unroll
         for (int b = 0; b < 4; ++b) {
             const int  k   = k_KQ * 4 + b;
@@ -163,7 +163,7 @@ static __device__ __forceinline__ float vec_dot_fattn_vec_KQ_q2_0(
             const uint8_t code = (K_q2_0[ib].qs[by] >> (2 * sub)) & 0x03;
             const float val = m + d * kQ2_0_lm_centroids[code];
             const int   qv  = (int)(int8_t)(u >> (8 * b));
-            const float q8eff = q8scale * qv - q8off;
+            const float q8eff = q8scale * (float)qv - q8off;
             sum += val * q8eff;
         }
     }
